backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
openai==1.82.0
orjson==3.10.18
packaging==24.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
